pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Denis Shumeev", email = "denis-shumeev@yandex.ru"},
]
dependencies = [
    "fastapi>=0.104.1",
    "sqlalchemy>=2.0.23",
    "uvicorn[standard]>=0.24.0.post1",
    "pydantic-settings>=2.0.3",
    "sentry-sdk>=1.34.0",
    "asyncpg>=0.29.0",
    "fastapi-users[sqlalchemy]>=12.1.2",
    "jinja2>=3.1.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

